+++ postgres_declare/data_structures/grant_on.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import Sequence
 
 from postgres_declare.data_structures.privileges import Privilege
-from postgres_declare.entities.grantable import Grantable
+from postgres_declare.mixins.grantable import Grantable
 
 
 @dataclass
