+++ postgres_declare/entities/database_content.py
@@ -8,8 +8,8 @@ from sqlalchemy.orm import DeclarativeBase
 from postgres_declare.entities.database import Database
 from postgres_declare.entities.database_entity import DatabaseEntity
 from postgres_declare.entities.entity import Entity
-from postgres_declare.entities.grantable import GrantableTable
 from postgres_declare.entities.schema import Schema
+from postgres_declare.mixins.grantable import GrantableTable
 
 
 class DatabaseContent(DatabaseEntity):
