+++ tests/unit/test_grantable.py
@@ -7,9 +7,9 @@ from postgres_declare.base import Base
 from postgres_declare.data_structures.grant_to import GrantTo
 from postgres_declare.data_structures.privileges import Privilege
 from postgres_declare.entities.entity import Entity
-from postgres_declare.entities.grantable import Grantable
 from postgres_declare.entities.role import Role
 from postgres_declare.exceptions import EntityExistsError, InvalidPrivilegeError
+from postgres_declare.mixins.grantable import Grantable
 from tests.helpers import YieldFixture
